+++ game/content/ghplots/lancedev.py
@@ -85,6 +85,8 @@ class LMMissionPlot(LMPlot):
 
 # The actual plots...
 
+
+
 class Earth_RescueAndBiotech(LMMissionPlot):
     LABEL = "LANCEDEV"
     active = True