+++ game/ghdialogue/__init__.py
@@ -76,7 +76,7 @@ pbge.dialogue.GRAMMAR_BUILDER = build_grammar
 pbge.dialogue.STANDARD_REPLIES = harvest(ghreplies, pbge.dialogue.Reply)
 pbge.dialogue.STANDARD_OFFERS = harvest(ghoffers, pbge.dialogue.Offer)
 pbge.dialogue.GENERIC_OFFERS.append(ghoffers.GOODBYE)
-pbge.dialogue.GENERIC_OFFERS.append(ghoffers.CHAT)
+#pbge.dialogue.GENERIC_OFFERS.append(ghoffers.CHAT)
 
 HELLO_STARTER = pbge.dialogue.Cue(pbge.dialogue.ContextTag((context.HELLO,)))
 UNFAVORABLE_STARTER = pbge.dialogue.Cue(pbge.dialogue.ContextTag((context.UNFAVORABLE_HELLO,)))
@@ -187,6 +187,58 @@ class MatchingTagPartyReply(SkillBasedPartyReply):
             return " [{} + NPC: ]".format(self.needed_tag, self.npc_tag)
 
 
+class GearHeadConversation(pbge.dialogue.DynaConversation):
+    CDATA_TODAYS_RUMORS = "CDATA_TODAYS_RUMORS"
+    CDATA_RUMOR_EXPIRATION = "CDATA_RUMOR_EXPIRATION"
+
+    def __init__(self, camp: gears.GearHeadCampaign, *args, **kwargs):
+        self.rumors = list()
+        self._gathered_rumors = False
+        self.chat_limit = 0
+        super().__init__(camp, *args, **kwargs)
+
+    def _get_dialogue_data(self):
+        super()._get_dialogue_data()
+        if not self._gathered_rumors:
+            if self.npc and self.camp.is_not_lancemate(self.npc):
+                cdata = self.camp.get_campdata(self.npc)
+                all_rumors = self.npc_grammar.get("[News]", ())
+                current_rumors = cdata.get(self.CDATA_TODAYS_RUMORS, ())
+                rumor_expiration = cdata.get(self.CDATA_RUMOR_EXPIRATION, -1)
+                react = self.npc.get_reaction_score(self.pc, self.camp)
+                self.chat_limit = max(react//10 + random.randint(-2,2), 1)
+                if all_rumors and (rumor_expiration < self.camp.time or not current_rumors):
+                    num_rumors = max(react//5 + random.randint(0, 5) - random.randint(0, 5), random.randint(1, 5))
+                    current_rumors = random.sample(all_rumors, num_rumors)
+                    cdata[self.CDATA_TODAYS_RUMORS] = current_rumors
+                    cdata[self.CDATA_RUMOR_EXPIRATION] = self.camp.time
+
+                if current_rumors:
+                    self.rumors = [r for r in current_rumors if r in all_rumors]
+                    random.shuffle(self.rumors)
+
+                self._gathered_rumors = True
+
+        if self.rumors:
+            self.npc_offers.append(Offer(
+                '[chat_lead_in] {}'.format(self.rumors[0]),
+                context=ContextTag([context.CHAT,]), is_generic=True,
+                prefx=self._display_first_rumor
+            ))
+        elif self.chat_limit > 0:
+            self.npc_offers.append(Offer(
+                '[CHAT]',
+                context=ContextTag([context.CHAT,]), is_generic=True,
+                prefx=self._random_chat
+            ))
+
+    def _display_first_rumor(self, _camp):
+        self.rumors.pop(0)
+
+    def _random_chat(self, _camp):
+        self.chat_limit -= 1
+
+
 def start_conversation(camp: gears.GearHeadCampaign, pc, npc, cue=None):
     # If this NPC has no relationship with the PC, create that now.
     realnpc = npc.get_pilot()
@@ -204,7 +256,7 @@ def start_conversation(camp: gears.GearHeadCampaign, pc, npc, cue=None):
             cue = HELLO_STARTER
     cviz = ghdview.ConvoVisualizer(npc, camp, pc=pc)
     cviz.rollout()
-    convo = pbge.dialogue.DynaConversation(camp, realnpc, pc, cue, visualizer=cviz)
+    convo = GearHeadConversation(camp, realnpc, pc, cue, visualizer=cviz)
     convo.converse()
     if realnpc:
         realnpc.relationship.met_before = True